+++ src/Adapter/Doctrine/ORMAdapter.php
@@ -12,11 +12,13 @@ declare(strict_types=1);
 
 namespace Omines\DataTablesBundle\Adapter\Doctrine;
 
+use Doctrine\Common\Persistence\ManagerRegistry;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Query;
 use Doctrine\ORM\QueryBuilder;
 use Omines\DataTablesBundle\Adapter\AbstractAdapter;
 use Omines\DataTablesBundle\Adapter\AdapterQuery;
+use Omines\DataTablesBundle\Adapter\Doctrine\Event\ORMAdapterQueryEvent;
 use Omines\DataTablesBundle\Adapter\Doctrine\ORM\AutomaticQueryBuilder;
 use Omines\DataTablesBundle\Adapter\Doctrine\ORM\QueryBuilderProcessorInterface;
 use Omines\DataTablesBundle\Adapter\Doctrine\ORM\SearchCriteriaProvider;
@@ -24,7 +26,6 @@ use Omines\DataTablesBundle\Column\AbstractColumn;
 use Omines\DataTablesBundle\DataTableState;
 use Omines\DataTablesBundle\Exception\InvalidConfigurationException;
 use Omines\DataTablesBundle\Exception\MissingDependencyException;
-use Symfony\Bridge\Doctrine\RegistryInterface;
 use Symfony\Component\OptionsResolver\Options;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -36,7 +37,7 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
  */
 class ORMAdapter extends AbstractAdapter
 {
-    /** @var RegistryInterface */
+    /** @var ManagerRegistry */
     private $registry;
 
     /** @var EntityManager */
@@ -56,10 +57,8 @@ class ORMAdapter extends AbstractAdapter
 
     /**
      * DoctrineAdapter constructor.
-     *
-     * @param RegistryInterface|null $registry
      */
-    public function __construct(RegistryInterface $registry = null)
+    public function __construct(ManagerRegistry $registry = null)
     {
         if (null === $registry) {
             throw new MissingDependencyException('Install doctrine/doctrine-bundle to use the ORMAdapter');
@@ -101,9 +100,6 @@ class ORMAdapter extends AbstractAdapter
         $this->criteriaProcessors[] = $this->normalizeProcessor($processor);
     }
 
-    /**
-     * @param AdapterQuery $query
-     */
     protected function prepareQuery(AdapterQuery $query)
     {
         $state = $query->getState();
@@ -133,7 +129,6 @@ class ORMAdapter extends AbstractAdapter
     }
 
     /**
-     * @param AdapterQuery $query
      * @return array
      */
     protected function getAliases(AdapterQuery $query)
@@ -151,6 +146,10 @@ class ORMAdapter extends AbstractAdapter
         foreach ($builder->getDQLPart('join') as $joins) {
             /** @var Query\Expr\Join $join */
             foreach ($joins as $join) {
+                if (false === mb_strstr($join->getJoin(), '.')) {
+                    continue;
+                }
+
                 list($origin, $target) = explode('.', $join->getJoin());
 
                 $mapping = $aliases[$origin][1]->getAssociationMapping($target);
@@ -169,10 +168,6 @@ class ORMAdapter extends AbstractAdapter
         return $this->mapFieldToPropertyPath($column->getField(), $query->get('aliases'));
     }
 
-    /**
-     * @param AdapterQuery $query
-     * @return \Traversable
-     */
     protected function getResults(AdapterQuery $query): \Traversable
     {
         /** @var QueryBuilder $builder */
@@ -193,17 +188,18 @@ class ORMAdapter extends AbstractAdapter
             ;
         }
 
-        foreach ($builder->getQuery()->iterate([], $this->hydrationMode) as $result) {
-            yield $entity = $result[0];
+        $query = $builder->getQuery();
+        $event = new ORMAdapterQueryEvent($query);
+        $state->getDataTable()->getEventDispatcher()->dispatch($event, ORMAdapterEvents::PRE_QUERY);
+
+        foreach ($query->iterate([], $this->hydrationMode) as $result) {
+            yield $entity = array_values($result)[0];
             if (Query::HYDRATE_OBJECT === $this->hydrationMode) {
                 $this->manager->detach($entity);
             }
         }
     }
 
-    /**
-     * @param DataTableState $state
-     */
     protected function buildCriteria(QueryBuilder $queryBuilder, DataTableState $state)
     {
         foreach ($this->criteriaProcessors as $provider) {
@@ -211,10 +207,6 @@ class ORMAdapter extends AbstractAdapter
         }
     }
 
-    /**
-     * @param DataTableState $state
-     * @return QueryBuilder
-     */
     protected function createQueryBuilder(DataTableState $state): QueryBuilder
     {
         /** @var QueryBuilder $queryBuilder */
@@ -238,7 +230,7 @@ class ORMAdapter extends AbstractAdapter
 
         $qb->resetDQLPart('orderBy');
         $gb = $qb->getDQLPart('groupBy');
-        if (empty($gb) || !in_array($identifier, $gb, true)) {
+        if (empty($gb) || !$this->hasGroupByPart($identifier, $gb)) {
             $qb->select($qb->expr()->count($identifier));
 
             return (int) $qb->getQuery()->getSingleScalarResult();
@@ -251,8 +243,23 @@ class ORMAdapter extends AbstractAdapter
     }
 
     /**
+     * @param $identifier
+     * @param Query\Expr\GroupBy[] $gbList
+     * @return bool
+     */
+    protected function hasGroupByPart($identifier, array $gbList)
+    {
+        foreach ($gbList as $gb) {
+            if (in_array($identifier, $gb->getParts(), true)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
      * @param string $field
-     * @param array $aliases
      * @return string
      */
     private function mapFieldToPropertyPath($field, array $aliases = [])
@@ -261,10 +268,13 @@ class ORMAdapter extends AbstractAdapter
         if (count($parts) < 2) {
             throw new InvalidConfigurationException(sprintf("Field name '%s' must consist at least of an alias and a field separated with a period", $field));
         }
-        list($origin, $target) = $parts;
 
-        $path = [$target];
-        $current = $aliases[$origin][0];
+        $origin = $parts[0];
+        array_shift($parts);
+        $target = array_reverse($parts);
+        $path = $target;
+
+        $current = isset($aliases[$origin]) ? $aliases[$origin][0] : null;
 
         while (null !== $current) {
             list($origin, $target) = explode('.', $current);
@@ -279,9 +289,6 @@ class ORMAdapter extends AbstractAdapter
         }
     }
 
-    /**
-     * @param OptionsResolver $resolver
-     */
     protected function configureOptions(OptionsResolver $resolver)
     {
         $providerNormalizer = function (Options $options, $value) {