+++ src/DataTableFactory.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 namespace Omines\DataTablesBundle;
 
 use Omines\DataTablesBundle\DependencyInjection\Instantiator;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\HttpFoundation\Request;
 
 class DataTableFactory
@@ -29,28 +30,28 @@ class DataTableFactory
     /** @var array */
     protected $config;
 
+    /** @var EventDispatcherInterface */
+    protected $eventDispatcher;
+
     /**
      * DataTableFactory constructor.
-     *
-     * @param array $config
-     * @param DataTableRendererInterface $renderer
      */
-    public function __construct(array $config, DataTableRendererInterface $renderer, Instantiator $instantiator)
+    public function __construct(array $config, DataTableRendererInterface $renderer, Instantiator $instantiator, EventDispatcherInterface $eventDispatcher)
     {
         $this->config = $config;
         $this->renderer = $renderer;
         $this->instantiator = $instantiator;
+        $this->eventDispatcher = $eventDispatcher;
     }
 
     /**
-     * @param array $options
      * @return DataTable
      */
     public function create(array $options = [])
     {
         $config = $this->config;
 
-        return (new DataTable(array_merge($config['options'] ?? [], $options), $this->instantiator))
+        return (new DataTable($this->eventDispatcher, array_merge($config['options'] ?? [], $options), $this->instantiator))
             ->setRenderer($this->renderer)
             ->setMethod($config['method'] ?? Request::METHOD_POST)
             ->setPersistState($config['persist_state'] ?? 'fragment')
@@ -62,8 +63,6 @@ class DataTableFactory
 
     /**
      * @param string|DataTableTypeInterface $type
-     * @param array $typeOptions
-     * @param array $options
      * @return DataTable
      */
     public function createFromType($type, array $typeOptions = [], array $options = [])