+++ src/Resources/public/js/datatables.js
@@ -13,6 +13,10 @@
      * Initializes the datatable dynamically.
      */
     $.fn.initDataTables = function(config, options) {
+
+        //Update default used url, so it reflects the current location (useful on single side apps)
+        $.fn.initDataTables.defaults.url = window.location.origin + window.location.pathname;
+
         var root = this,
             config = $.extend({}, $.fn.initDataTables.defaults, config),
             state = ''
@@ -55,8 +59,10 @@
                             drawCallback(data);
                             data = null;
                             if (Object.keys(state).length) {
-                                var merged = $.extend(true, {}, dt.state(), state);
-                                dt
+                                var api = new $.fn.dataTable.Api( settings );
+                                var merged = $.extend(true, {}, api.state(), state);
+
+                                api
                                     .order(merged.order)
                                     .search(merged.search.search)
                                     .page.len(merged.length)
@@ -88,10 +94,12 @@
                             var diff = data.filter(el => { return baseState.indexOf(el) === -1 && el.indexOf('time=') !== 0; });
                             switch (config.state) {
                                 case 'fragment':
-                                    history.replaceState(null, null, '#' + decodeURIComponent(diff.join('&')));
+                                    history.replaceState(null, null, window.location.origin + window.location.pathname + window.location.search
+                                        + '#' + decodeURIComponent(diff.join('&')));
                                     break;
                                 case 'query':
-                                    history.replaceState(null, null, '?' + decodeURIComponent(diff.join('&')));
+                                    history.replaceState(null, null, window.location.origin + window.location.pathname
+                                        + '?' + decodeURIComponent(diff.join('&') + window.location.hash));
                                     break;
                             }
                         }