+++ tests/Fixtures/AppBundle/Controller/GroupedController.php
@@ -12,9 +12,10 @@ declare(strict_types=1);
 
 namespace Tests\Fixtures\AppBundle\Controller;
 
-use Omines\DataTablesBundle\Controller\DataTablesTrait;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Omines\DataTablesBundle\DataTableFactory;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
+use Tests\Fixtures\AppBundle\DataTable\Type\Grouped2TableType;
 use Tests\Fixtures\AppBundle\DataTable\Type\GroupedTableType;
 
 /**
@@ -22,14 +23,22 @@ use Tests\Fixtures\AppBundle\DataTable\Type\GroupedTableType;
  *
  * @author Niels Keurentjes <niels.keurentjes@omines.com>
  */
-class GroupedController extends Controller
+class GroupedController extends AbstractController
 {
-    use DataTablesTrait;
-
-    public function tableAction(Request $request)
+    public function tableAction(Request $request, DataTableFactory $dataTableFactory)
     {
-        $datatable = $this->createDataTableFromType(GroupedTableType::class)
+        $datatable = $dataTableFactory->createFromType(GroupedTableType::class)
             ->setName('companies')
+            ->setMethod(Request::METHOD_GET)
+        ;
+
+        return $datatable->handleRequest($request)->getResponse();
+    }
+
+    public function table2Action(Request $request, DataTableFactory $dataTableFactory)
+    {
+        $datatable = $dataTableFactory->createFromType(Grouped2TableType::class)
+            ->setName('companies2')
             ->setMethod(Request::METHOD_GET)
         ;