+++ tests/Fixtures/AppBundle/Controller/ServiceController.php
@@ -12,8 +12,8 @@ declare(strict_types=1);
 
 namespace Tests\Fixtures\AppBundle\Controller;
 
-use Omines\DataTablesBundle\Controller\DataTablesTrait;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Omines\DataTablesBundle\DataTableFactory;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Tests\Fixtures\AppBundle\DataTable\Type\ServicePersonTableType;
 
@@ -22,13 +22,11 @@ use Tests\Fixtures\AppBundle\DataTable\Type\ServicePersonTableType;
  *
  * @author Niels Keurentjes <niels.keurentjes@omines.com>
  */
-class ServiceController extends Controller
+class ServiceController extends AbstractController
 {
-    use DataTablesTrait;
-
-    public function tableAction(Request $request)
+    public function tableAction(Request $request, DataTableFactory $dataTableFactory)
     {
-        $datatable = $this->createDataTableFromType(ServicePersonTableType::class, [], ['order' => [[1, 'asc']]])
+        $datatable = $dataTableFactory->createFromType(ServicePersonTableType::class, [], ['order' => [[1, 'asc']]])
             ->setName('persons')
             ->setMethod(Request::METHOD_GET)
         ;