+++ tests/Unit/ColumnTest.php
@@ -15,10 +15,12 @@ namespace Tests\Unit;
 use Omines\DataTablesBundle\Column\BoolColumn;
 use Omines\DataTablesBundle\Column\DateTimeColumn;
 use Omines\DataTablesBundle\Column\MapColumn;
+use Omines\DataTablesBundle\Column\NumberColumn;
 use Omines\DataTablesBundle\Column\TextColumn;
 use Omines\DataTablesBundle\Column\TwigColumn;
 use Omines\DataTablesBundle\DataTable;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\EventDispatcher\EventDispatcher;
 
 /**
  * ColumnTest.
@@ -33,7 +35,7 @@ class ColumnTest extends TestCase
         $column->initialize('test', 1, [
             'nullValue' => 'foo',
             'format' => 'd-m-Y',
-        ], (new DataTable())->setName('foo'));
+        ], (new DataTable($this->createMock(EventDispatcher::class)))->setName('foo'));
 
         $this->assertSame('03-04-2015', $column->transform('2015-04-03'));
         $this->assertSame('foo', $column->transform(null));
@@ -45,7 +47,7 @@ class ColumnTest extends TestCase
         $column->initialize('test', 1, [
             'data' => 'bar',
             'render' => 'foo%s',
-        ], (new DataTable())->setName('foo'));
+        ], (new DataTable($this->createMock(EventDispatcher::class)))->setName('foo'));
 
         $this->assertFalse($column->isRaw());
         $this->assertSame('foobar', $column->transform(null));
@@ -58,12 +60,18 @@ class ColumnTest extends TestCase
         $column->initialize('test', 1, [
              'trueValue' => 'yes',
              'nullValue' => '<em>null</em>',
-        ], new DataTable());
+        ], new DataTable($this->createMock(EventDispatcher::class)));
 
         $this->assertSame('yes', $column->transform(5));
         $this->assertSame('yes', $column->transform(true));
         $this->assertSame('false', $column->transform(false));
         $this->assertStringStartsWith('<em>', $column->transform());
+
+        $this->assertTrue($column->isValidForSearch('yes'));
+        $this->assertFalse($column->isValidForSearch('true'));
+
+        $this->assertTrue($column->getRightExpr('yes'));
+        $this->assertFalse($column->getRightExpr('true'));
     }
 
     public function testMapColumn()
@@ -75,12 +83,26 @@ class ColumnTest extends TestCase
                 1 => 'bar',
                 2 => 'baz',
             ],
-        ], new DataTable());
+        ], new DataTable($this->createMock(EventDispatcher::class)));
 
         $this->assertSame('foo', $column->transform(0));
         $this->assertSame('bar', $column->transform(1));
         $this->assertSame('baz', $column->transform(2));
         $this->assertSame('foo', $column->transform(3));
+    }
+
+    public function testNumberColumn()
+    {
+        $column = new NumberColumn();
+        $column->initialize('test', 1, [], new DataTable($this->createMock(EventDispatcher::class)));
+
+        $this->assertSame('5', $column->transform(5));
+        $this->assertSame('1', $column->transform(true));
+        $this->assertSame('684', $column->transform('684'));
+
+        $this->assertFalse($column->isRaw());
+        $this->assertTrue($column->isValidForSearch(684));
+        $this->assertFalse($column->isValidForSearch('foo.bar'));
     }
 
     public function testColumnWithClosures()
@@ -93,7 +115,7 @@ class ColumnTest extends TestCase
             'render' => function ($value) {
                 return mb_strtoupper($value);
             },
-        ], new DataTable());
+        ], new DataTable($this->createMock(EventDispatcher::class)));
 
         $this->assertFalse($column->isRaw());
         $this->assertSame('BAR', $column->transform(null));